- tooltips for TopToolbar buttons - replaceStr/replaceStrW moved to mir_core
git-svn-id: http://svn.miranda-ng.org/main/trunk@886 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Clist_modern/modern_newrowopts.cpp
@@ -74,7 +74,7 @@ BOOL RemoveChildNode(NodeList * FromList, DWORD index)
 	NodeList *work = &(FromList->childNodes[index]);
 	for (size_t i=0; i < work->AllocatedChilds; i++)
 		if (work->childNodes[i].AllocatedChilds)
-			RemoveChildNode(work->childNodes,i);
+			RemoveChildNode(work->childNodes, (DWORD)i);
 
 	if (work->AllocatedChilds) {
 		mir_free_and_nil(work->childNodes);
